Code smell Remove this unnecessary cast to 'long'.

+++ src/SME.SGP.Dados/Repositorios/RepositorioComunicadoAnoEscolar.cs
@@ -20,7 +20,7 @@ namespace SME.SGP.Dados.Repositorios
             if (comunicadoAnoEscolar.Id > 0)
                 await database.Conexao.UpdateMappedAsync(comunicadoAnoEscolar);
             else
-                comunicadoAnoEscolar.Id = (long)(await database.Conexao.InsertMappedAsync(comunicadoAnoEscolar));
+                comunicadoAnoEscolar.Id = (await database.Conexao.InsertMappedAsync(comunicadoAnoEscolar));
 
             return comunicadoAnoEscolar.Id;
         }
