+++ src/SME.SGP.Dados/Repositorios/RepositorioOcorrenciaAluno.cs
@@ -40,7 +40,7 @@ namespace SME.SGP.Dados
             }
             else
             {
-                entidade.Id = (long)(await database.Conexao.InsertMappedAsync(entidade));
+                entidade.Id = (await database.Conexao.InsertMappedAsync(entidade));
                 await AuditarAsync(entidade.Id, "I");
             }
